除了未完成部分，测试都已经通过，详细情况见聊天记录

+++ src/main/java/com/mobileai/dxc/service/imple/ProvisionServiceImple.java
@@ -7,6 +7,7 @@ import com.mobileai.dxc.db.pojo.Picture;
 import com.mobileai.dxc.db.pojo.Provision;
 import com.mobileai.dxc.db.pojo.Seller;
 import com.mobileai.dxc.service.ProvisionService;
+import com.mobileai.dxc.util.Result;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -19,6 +20,7 @@ import static com.mobileai.dxc.util.DeleteDirectory.deleteDirectory;
 import static com.mobileai.dxc.util.ImageUtil.generateThumbnail;
 import static com.mobileai.dxc.util.IntStringUtils.addSubString;
 import static com.mobileai.dxc.util.IntStringUtils.deleteSubString;
+import static com.mobileai.dxc.util.IntStringUtils.replaceString;
 import static com.mobileai.dxc.util.PathUtil.getImgBasePath;
 
 @Service
@@ -35,7 +37,8 @@ public class ProvisionServiceImple implements ProvisionService {
 
 
     @Override
-    public int addSerive(int sellerId,String  name,String description,float price,float prePrice) {
+    @Transactional
+    public int addService(int sellerId, String  name, String description, float price, float prePrice) {
         Provision provision = new Provision();
         provision.setSellerId(sellerId);
         provision.setName(name);
@@ -64,14 +67,18 @@ public class ProvisionServiceImple implements ProvisionService {
 
     @Override
     @Transactional
-    public boolean deleteService(int sellerId, int serviceId) {
+    public Result deleteService(int sellerId, int serviceId) {
 
         int status =serviceMapper.selectStatusById(serviceId);
-        if(status ==0)
-        {throw  new RuntimeException("该商品正在审核,商品不能下架");}
+
+        //商品已经下架
         if(status ==2)
-        {throw  new RuntimeException("该商品已经下架");}
-        if(status ==1)
+        {
+            return new Result(100,"商品已经下架",null);
+        }
+
+        //商品状态正常1 商品在审核中0
+        if(status ==1||status ==0)
         {
             Provision provision =new Provision();
             provision.setServiceId(serviceId);
@@ -82,11 +89,21 @@ public class ProvisionServiceImple implements ProvisionService {
             if(orderNum ==0)//todo:每次服务后，odernum-1后检查状态是下架，就删除服务
             {
                 String  pictureDirPath =serviceMapper.selectPictureById(serviceId);
-                System.out.println("pictureFilePath"+pictureDirPath);
-                deleteDirectory(pictureDirPath);
+                if(pictureDirPath != null)
+                {
+                    deleteDirectory(pictureDirPath);
+                }
 
-                int effcetNum= serviceMapper.deleteById(serviceId);
-                if(effcetNum<1) throw new RuntimeException("删除服务失败");
+                try{
+                    int effectNum= serviceMapper.deleteById(serviceId);
+                    if(effectNum < 1)
+                    {
+                        throw  new Exception("服务删除失败");
+                    }
+                }catch(Exception e)
+                {
+                    return  new Result(100,"删除服务失败，可联系维护人员进行咨询",null);
+                }
 
                 String sellerService =sellerMapper.selectServiceById(sellerId);
                 sellerService =deleteSubString(serviceId+"",sellerService);
@@ -94,43 +111,60 @@ public class ProvisionServiceImple implements ProvisionService {
                 seller.setSellerId(sellerId);
                 seller.setService(sellerService);
                 sellerMapper.updateSellerById(seller);
-
+                return  new Result(200,"服务删除成功",null);
             }
+            return  new Result(100,"该商品有订单未处理","未处理商品数为"+serviceMapper.selectOrderNumById(serviceId));
         }
-        return  false;
+        else return  new Result(100,"商品状态异常，请联系维护人员进行咨询");
     }
 
     @Override
-        public boolean updateService(int serviceId, Provision provision) {
+    @Transactional
+    public Result updateService(int sellerId, int serviceId, Provision provision) {
 
         provision.setServiceId(serviceId);
-        serviceMapper.updateServiceById(provision);
+        if(sellerId != serviceMapper.selectSellerIdByserviceId(serviceId) )
+        {
+            return new Result(100,"更新服务，参数传输错误",null);
+        }
 
-        return false;
+       int effectNum = serviceMapper.updateServiceById(provision);
+        //test :是影响行数 还是id
+        System.out.println("effectNum"+effectNum);
+
+        if(effectNum < 1){ return  new Result(100,"更新服务失败",null);}
+        else{return new Result(200,"更新服务成功",serviceMapper.getServiceById(serviceId));}
     }
 
 
     @Override
-    public boolean deleteServicePicture(int serviceId, int pictureId) {
+    @Transactional
+    public Result deleteServicePicture(int serviceId, int pictureId) {
         //确认id一致
         Picture picture=pictureMapper.selectById(pictureId);
         if(serviceId != picture.getServiceId())
         {
-            throw  new RuntimeException("该服务无此照片！");
+            return new Result(100,"传入数据错误",null);
         }
         //删除照片
-        File pictureFile =new File(picture.getPicturePath());
+        String dirPath =serviceMapper.selectPictureById(serviceId);
+        File pictureFile =new File(dirPath+System.getProperty("file.separator")+picture.getPicturePath());
+
+        System.out.println("file"+dirPath+System.getProperty("file.separator")+picture.getPicturePath());
         pictureFile.delete();
 
-       int effetNum = pictureMapper.deletePictureByid(pictureId);
-        if(effetNum < 1)
+       int effectNum = pictureMapper.deletePictureByid(pictureId);
+        if(effectNum < 1)
         {
-            throw  new RuntimeException("删除照片失败！");
+            return new Result(100,"删除照片失败",null);
         }
-        return false;
+        else{
+            System.out.println("goes here");
+            return  new Result(200,"删除照片成功",null); }
     }
 
     @Override
+    @Transactional
     public String addServicePicture(int serviceId, MultipartFile thumbnai) {
         //把照片存储到合适的位置
         Provision provision =serviceMapper.getServiceById(serviceId);
@@ -151,17 +185,20 @@ public class ProvisionServiceImple implements ProvisionService {
     }
 
     @Override
+    @Transactional
     public List<Picture>  selectAllServicePicture(int serviceId)
     {
         return pictureMapper.selectPictureByServiceId(serviceId);
     }
 
     @Override
+    @Transactional
     public List<Provision> showAllService(int sellerId) {
         return serviceMapper.selectServiceBySellerId(sellerId);
     }
 
     @Override
+    @Transactional
     public List<Provision> findPreferService() {
         return null;
     }
